Replace follow point renderer with new implementation
