+++ osu.Game.Rulesets.Osu/Objects/Drawables/Connections/FollowPointRenderer.cs
@@ -1,122 +1,104 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Collections.Generic;
-using osuTK;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Game.Rulesets.Objects.Types;
 
 namespace osu.Game.Rulesets.Osu.Objects.Drawables.Connections
 {
     public class FollowPointRenderer : CompositeDrawable
     {
-        private int pointDistance = 32;
+        /// <summary>
+        /// Adds the <see cref="FollowPoint"/>s around a <see cref="DrawableOsuHitObject"/>.
+        /// This includes <see cref="FollowPoint"/>s leading into <paramref name="hitObject"/>, and <see cref="FollowPoint"/>s exiting <paramref name="hitObject"/>.
+        /// </summary>
+        /// <param name="hitObject">The <see cref="DrawableOsuHitObject"/> to add <see cref="FollowPoint"/>s for.</param>
+        public void AddFollowPoints(DrawableOsuHitObject hitObject)
+        {
+            var startGroup = new FollowPointGroup(hitObject);
+            AddInternal(startGroup);
+
+            // Groups are sorted by their start time when added, so the index can be used to post-process other surrounding groups
+            int startIndex = IndexOfInternal(startGroup);
+
+            if (startIndex < InternalChildren.Count - 1)
+            {
+                //     h1 -> -> -> h2
+                //  hitObject   nextGroup
+
+                var nextGroup = (FollowPointGroup)InternalChildren[startIndex + 1];
+                startGroup.End = nextGroup.Start;
+            }
+
+            if (startIndex > 0)
+            {
+                //     h1 -> -> -> h2
+                //  prevGroup   hitObject
+
+                var previousGroup = (FollowPointGroup)InternalChildren[startIndex - 1];
+                previousGroup.End = startGroup.Start;
+            }
+        }
 
         /// <summary>
-        /// Determines how much space there is between points.
+        /// Removes the <see cref="FollowPoint"/>s around a <see cref="DrawableOsuHitObject"/>.
+        /// This includes <see cref="FollowPoint"/>s leading into <paramref name="hitObject"/>, and <see cref="FollowPoint"/>s exiting <paramref name="hitObject"/>.
         /// </summary>
-        public int PointDistance
+        /// <param name="hitObject">The <see cref="DrawableOsuHitObject"/> to remove <see cref="FollowPoint"/>s for.</param>
+        public void RemoveFollowPoints(DrawableOsuHitObject hitObject)
         {
-            get => pointDistance;
-            set
-            {
-                if (pointDistance == value) return;
+            var groups = findGroups(hitObject);
 
-                pointDistance = value;
-                update();
+            // Regardless of the position of the hitobject in the beatmap, there will always be a group leading from the hitobject
+            RemoveInternal(groups.start);
+
+            if (groups.end != null)
+            {
+                // When there were two groups referencing the same hitobject,  merge them by updating the end group to point to the new end (the start group was already removed)
+                groups.end.End = groups.start.End;
             }
         }
-
-        private int preEmpt = 800;
 
         /// <summary>
-        /// Follow points to the next hitobject start appearing for this many milliseconds before an hitobject's end time.
+        /// Finds the <see cref="FollowPointGroup"/>s with <paramref name="hitObject"/> as the start and end <see cref="DrawableOsuHitObject"/>s.
         /// </summary>
-        public int PreEmpt
+        /// <param name="hitObject">The <see cref="DrawableOsuHitObject"/> to find the relevant <see cref="FollowPointGroup"/> of.</param>
+        /// <returns>A tuple containing the end group (the <see cref="FollowPointGroup"/> where <paramref name="hitObject"/> is the end of),
+        /// and the start group (the <see cref="FollowPointGroup"/> where <paramref name="hitObject"/> is the start of).</returns>
+        private (FollowPointGroup start, FollowPointGroup end) findGroups(DrawableOsuHitObject hitObject)
         {
-            get => preEmpt;
-            set
+            //           endGroup         startGroup
+            //     h1 -> -> -> -> -> h2 -> -> -> -> -> h3
+            //                    hitObject
+
+            FollowPointGroup startGroup = null; // The group which the hitobject is the start in
+            FollowPointGroup endGroup = null; // The group which the hitobject is the end in
+
+            int startIndex = 0;
+
+            for (; startIndex < InternalChildren.Count; startIndex++)
             {
-                if (preEmpt == value) return;
+                var group = (FollowPointGroup)InternalChildren[startIndex];
 
-                preEmpt = value;
-                update();
-            }
-        }
-
-        private IEnumerable<OsuHitObject> hitObjects;
-
-        public IEnumerable<OsuHitObject> HitObjects
-        {
-            get => hitObjects;
-            set
-            {
-                hitObjects = value;
-                update();
-            }
-        }
-
-        public override bool RemoveCompletedTransforms => false;
-
-        private void update()
-        {
-            ClearInternal();
-
-            if (hitObjects == null)
-                return;
-
-            OsuHitObject prevHitObject = null;
-
-            foreach (var currHitObject in hitObjects)
-            {
-                if (prevHitObject != null && !currHitObject.NewCombo && !(prevHitObject is Spinner) && !(currHitObject is Spinner))
+                if (group.Start == hitObject)
                 {
-                    Vector2 startPosition = prevHitObject.EndPosition;
-                    Vector2 endPosition = currHitObject.Position;
-                    double startTime = (prevHitObject as IHasEndTime)?.EndTime ?? prevHitObject.StartTime;
-                    double endTime = currHitObject.StartTime;
-
-                    Vector2 distanceVector = endPosition - startPosition;
-                    int distance = (int)distanceVector.Length;
-                    float rotation = (float)(Math.Atan2(distanceVector.Y, distanceVector.X) * (180 / Math.PI));
-                    double duration = endTime - startTime;
-
-                    for (int d = (int)(PointDistance * 1.5); d < distance - PointDistance; d += PointDistance)
-                    {
-                        float fraction = (float)d / distance;
-                        Vector2 pointStartPosition = startPosition + (fraction - 0.1f) * distanceVector;
-                        Vector2 pointEndPosition = startPosition + fraction * distanceVector;
-                        double fadeOutTime = startTime + fraction * duration;
-                        double fadeInTime = fadeOutTime - PreEmpt;
-
-                        FollowPoint fp;
-
-                        AddInternal(fp = new FollowPoint
-                        {
-                            Position = pointStartPosition,
-                            Rotation = rotation,
-                            Alpha = 0,
-                            Scale = new Vector2(1.5f * currHitObject.Scale),
-                        });
-
-                        using (fp.BeginAbsoluteSequence(fadeInTime))
-                        {
-                            fp.FadeIn(currHitObject.TimeFadeIn);
-                            fp.ScaleTo(currHitObject.Scale, currHitObject.TimeFadeIn, Easing.Out);
-
-                            fp.MoveTo(pointEndPosition, currHitObject.TimeFadeIn, Easing.Out);
-
-                            fp.Delay(fadeOutTime - fadeInTime).FadeOut(currHitObject.TimeFadeIn);
-                        }
-
-                        fp.Expire(true);
-                    }
+                    startGroup = group;
+                    break;
                 }
-
-                prevHitObject = currHitObject;
             }
+
+            if (startIndex > 0)
+                endGroup = (FollowPointGroup)InternalChildren[startIndex - 1];
+
+            return (startGroup, endGroup);
+        }
+
+        protected override int Compare(Drawable x, Drawable y)
+        {
+            var groupX = (FollowPointGroup)x;
+            var groupY = (FollowPointGroup)y;
+
+            return groupX.Start.HitObject.StartTime.CompareTo(groupY.Start.HitObject.StartTime);
         }
     }
 }
